[Style] #77 - 컴포넌트 폴더 생성 및 수정
